separated annotation builder and attribute builder into separate classes
