+++ src/Exception/IncompatiblePhpVersionException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Laminas\Form\Exception;
+
+class IncompatiblePhpVersionException extends \RuntimeException implements
+    ExceptionInterface
+{
+}
