+++ test/Annotation/AnnotationBuilderTest.php
@@ -3,523 +3,11 @@
 namespace LaminasTest\Form\Annotation;
 
 use Laminas\Form\Annotation;
-use Laminas\Hydrator\ClassMethods;
-use Laminas\Hydrator\ClassMethodsHydrator;
-use Laminas\Hydrator\ObjectProperty;
-use Laminas\Hydrator\ObjectPropertyHydrator;
-use LaminasTest\Form\TestAsset;
-use PHPUnit\Framework\TestCase;
 
-use function class_exists;
-use function getenv;
-
-class AnnotationBuilderTest extends TestCase
+class AnnotationBuilderTest extends AbstractBuilderTestCase
 {
-    /** @var string */
-    private $classMethodsHydratorClass;
-
-    /** @var string */
-    private $objectPropertyHydratorClass;
-
-    protected function setUp(): void
+    protected function createBuilder(): Annotation\AbstractBuilder
     {
-        if (! getenv('TESTS_LAMINAS_FORM_ANNOTATION_SUPPORT')) {
-            $this->markTestSkipped('Enable TESTS_LAMINAS_FORM_ANNOTATION_SUPPORT to test annotation parsing');
-        }
-
-        $this->classMethodsHydratorClass = class_exists(ClassMethodsHydrator::class)
-            ? ClassMethodsHydrator::class
-            : ClassMethods::class;
-
-        $this->objectPropertyHydratorClass = class_exists(ObjectPropertyHydrator::class)
-            ? ObjectPropertyHydrator::class
-            : ObjectProperty::class;
-    }
-
-    /**
-     * @return \Generator
-     */
-    public function useAttributesDataProvider()
-    {
-        if (PHP_MAJOR_VERSION >= 8) {
-            yield [true];
-        }
-
-        yield [false];
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testCanCreateFormFromStandardEntity($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\Entity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertTrue($form->has('username'));
-        $this->assertTrue($form->has('password'));
-
-        $username = $form->get('username');
-        $this->assertInstanceOf('Laminas\Form\Element', $username);
-        $this->assertEquals('required', $username->getAttribute('required'));
-
-        $password = $form->get('password');
-        $this->assertInstanceOf('Laminas\Form\Element', $password);
-        $attributes = $password->getAttributes();
-        $this->assertEquals(
-            ['type' => 'password', 'label' => 'Enter your password', 'name' => 'password'],
-            $attributes
-        );
-        $this->assertNull($password->getAttribute('required'));
-
-        $filter = $form->getInputFilter();
-        $this->assertTrue($filter->has('username'));
-        $this->assertTrue($filter->has('password'));
-
-        $username = $filter->get('username');
-        $this->assertTrue($username->isRequired());
-        $this->assertCount(1, $username->getFilterChain());
-        $this->assertCount(2, $username->getValidatorChain());
-
-        $password = $filter->get('password');
-        $this->assertTrue($password->isRequired());
-        $this->assertCount(1, $password->getFilterChain());
-        $this->assertCount(1, $password->getValidatorChain());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testCanCreateFormWithClassAnnotations($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\ClassEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertTrue($form->has('keeper'));
-        $this->assertFalse($form->has('keep'));
-        $this->assertFalse($form->has('omit'));
-        $this->assertEquals('some_name', $form->getName());
-
-        $attributes = $form->getAttributes();
-        $this->assertArrayHasKey('legend', $attributes);
-        $this->assertEquals('Some Fieldset', $attributes['legend']);
-
-        $filter = $form->getInputFilter();
-        $this->assertInstanceOf('LaminasTest\Form\TestAsset\Annotation\InputFilter', $filter);
-
-        $keeper     = $form->get('keeper');
-        $attributes = $keeper->getAttributes();
-        $this->assertArrayHasKey('type', $attributes);
-        $this->assertEquals('text', $attributes['type']);
-
-        $this->assertSame(['omit', 'keep'], $form->getValidationGroup());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testComplexEntityCreationWithPriorities($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\ComplexEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertEquals('user', $form->getName());
-        $attributes = $form->getAttributes();
-        $this->assertArrayHasKey('legend', $attributes);
-        $this->assertEquals('Register', $attributes['legend']);
-
-        $this->assertFalse($form->has('someComposedObject'));
-        $this->assertTrue($form->has('user_image'));
-        $this->assertTrue($form->has('email'));
-        $this->assertTrue($form->has('password'));
-        $this->assertTrue($form->has('username'));
-
-        $email = $form->get('email');
-        $test  = $form->getIterator()->getIterator()->current();
-        $this->assertSame($email, $test, 'Test is element ' . $test->getName());
-
-        $test  = $form->getIterator()->current();
-        $this->assertSame($email, $test, 'Test is element ' . $test->getName());
-
-        $hydrator = $form->getHydrator();
-        $this->assertInstanceOf($this->objectPropertyHydratorClass, $hydrator);
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testFieldsetOrder($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\FieldsetOrderEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $element = $form->get('element');
-        $first  = $form->getIterator()->getIterator()->current();
-        $this->assertSame($element, $first, 'Test is element ' . $first->getName());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testFieldsetOrderWithPreserve($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\FieldsetOrderEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $builder->setPreserveDefinedOrder(true);
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $fieldset = $form->get('fieldset');
-        $first  = $form->getIterator()->getIterator()->current();
-        $this->assertSame($fieldset, $first, 'Test is element ' . $first->getName());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testCanRetrieveOnlyFormSpecification($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\ComplexEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $spec    = $builder->getFormSpecification($entity, $useAttributes);
-        $this->assertInstanceOf('ArrayObject', $spec);
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testAllowsExtensionOfEntities($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\ExtendedEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertTrue($form->has('username'));
-        $this->assertTrue($form->has('password'));
-        $this->assertTrue($form->has('email'));
-
-        $this->assertEquals('extended', $form->getName());
-        $expected = ['username', 'password', 'email'];
-        $test     = [];
-        foreach ($form as $element) {
-            $test[] = $element->getName();
-        }
-        $this->assertEquals($expected, $test);
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testAllowsSpecifyingFormAndElementTypes($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\TypedEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertInstanceOf('LaminasTest\Form\TestAsset\Annotation\Form', $form);
-        $element = $form->get('typed_element');
-        $this->assertInstanceOf('LaminasTest\Form\TestAsset\Annotation\Element', $element);
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testAllowsComposingChildEntities($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\EntityComposingAnEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertTrue($form->has('composed'));
-        $composed = $form->get('composed');
-        $this->assertInstanceOf('Laminas\Form\FieldsetInterface', $composed);
-        $this->assertTrue($composed->has('username'));
-        $this->assertTrue($composed->has('password'));
-
-        $filter = $form->getInputFilter();
-        $this->assertTrue($filter->has('composed'));
-        $composed = $filter->get('composed');
-        $this->assertInstanceOf('Laminas\InputFilter\InputFilterInterface', $composed);
-        $this->assertTrue($composed->has('username'));
-        $this->assertTrue($composed->has('password'));
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testAllowsComposingMultipleChildEntities($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\EntityComposingMultipleEntities();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertTrue($form->has('composed'));
-        $composed = $form->get('composed');
-
-        $this->assertInstanceOf('Laminas\Form\Element\Collection', $composed);
-        $target = $composed->getTargetElement();
-        $this->assertInstanceOf('Laminas\Form\FieldsetInterface', $target);
-        $this->assertTrue($target->has('username'));
-        $this->assertTrue($target->has('password'));
-    }
-
-    /**
-     * @dataProvider provideOptionsAnnotationAndComposedObjectAnnotation
-     * @param string $childName
-     *
-     * @group issue-7108
-     */
-    public function testOptionsAnnotationAndComposedObjectAnnotation($useAttributes, $childName)
-    {
-        $entity  = new TestAsset\Annotation\EntityUsingComposedObjectAndOptions();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $child = $form->get($childName);
-
-        $target = $child->getTargetElement();
-        $this->assertInstanceOf('Laminas\Form\FieldsetInterface', $target);
-        $this->assertEquals('My label', $child->getLabel());
-    }
-
-    /**
-     * Data provider
-     *
-     * @return \Generator
-     */
-    public function provideOptionsAnnotationAndComposedObjectAnnotation()
-    {
-        if (PHP_MAJOR_VERSION >= 8) {
-            yield [true, 'child'];
-            yield [true, 'childTheSecond'];
-        }
-
-        yield [false, 'child'];
-        yield [false, 'childTheSecond'];
-    }
-
-    /**
-     * @dataProvider provideOptionsAnnotationAndComposedObjectAnnotationNoneCollection
-     * @param string $childName
-     *
-     * @group issue-7108
-     */
-    public function testOptionsAnnotationAndComposedObjectAnnotationNoneCollection($useAttributes, $childName)
-    {
-        $entity  = new TestAsset\Annotation\EntityUsingComposedObjectAndOptions();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $child = $form->get($childName);
-
-        $this->assertInstanceOf('Laminas\Form\FieldsetInterface', $child);
-        $this->assertEquals('My label', $child->getLabel());
-    }
-
-    /**
-     * Data provider
-     *
-     * @return \Generator
-     */
-    public function provideOptionsAnnotationAndComposedObjectAnnotationNoneCollection()
-    {
-        if (PHP_MAJOR_VERSION >= 8) {
-            yield [true, 'childTheThird'];
-            yield [true, 'childTheFourth'];
-        }
-
-        yield [false, 'childTheThird'];
-        yield [false, 'childTheFourth'];
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testCanHandleOptionsAnnotation($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\EntityUsingOptions();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertTrue($form->useAsBaseFieldset());
-
-        $this->assertTrue($form->has('username'));
-
-        $username = $form->get('username');
-        $this->assertInstanceOf('Laminas\Form\Element', $username);
-
-        $this->assertEquals('Username:', $username->getLabel());
-        $this->assertEquals(['class' => 'label'], $username->getLabelAttributes());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testCanHandleHydratorArrayAnnotation($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\EntityWithHydratorArray();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $hydrator = $form->getHydrator();
-        $this->assertInstanceOf($this->classMethodsHydratorClass, $hydrator);
-        $this->assertFalse($hydrator->getUnderscoreSeparatedKeys());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testAllowTypeAsElementNameInInputFilter($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\EntityWithTypeAsElementName();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $this->assertInstanceOf('Laminas\Form\Form', $form);
-        $element = $form->get('type');
-        $this->assertInstanceOf('Laminas\Form\Element', $element);
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testAllowEmptyInput($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\SampleEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $inputFilter = $form->getInputFilter();
-        $sampleinput = $inputFilter->get('sampleinput');
-        $this->assertTrue($sampleinput->allowEmpty());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testContinueIfEmptyInput($useAttributes)
-    {
-        $entity  = new TestAsset\Annotation\SampleEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-
-        $inputFilter = $form->getInputFilter();
-        $sampleinput = $inputFilter->get('sampleinput');
-        $this->assertTrue($sampleinput->continueIfEmpty());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testInputNotRequiredByDefault($useAttributes)
-    {
-        $entity = new TestAsset\Annotation\SampleEntity();
-        $builder = new Annotation\AnnotationBuilder();
-        $form = $builder->createForm($entity, $useAttributes);
-        $inputFilter = $form->getInputFilter();
-        $sampleinput = $inputFilter->get('anotherSampleInput');
-        $this->assertFalse($sampleinput->isRequired());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testInstanceElementAnnotation($useAttributes)
-    {
-        $entity = new TestAsset\Annotation\EntityUsingInstanceProperty();
-        $builder = new Annotation\AnnotationBuilder();
-        $form = $builder->createForm($entity, $useAttributes);
-
-        $fieldset = $form->get('object');
-        /* @var $fieldset Laminas\Form\Fieldset */
-
-        $this->assertInstanceOf('Laminas\Form\Fieldset', $fieldset);
-        $this->assertInstanceOf('LaminasTest\Form\TestAsset\Annotation\Entity', $fieldset->getObject());
-        $this->assertInstanceOf($this->classMethodsHydratorClass, $fieldset->getHydrator());
-        $this->assertFalse($fieldset->getHydrator()->getUnderscoreSeparatedKeys());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testInputFilterInputAnnotation($useAttributes)
-    {
-        $entity = new TestAsset\Annotation\EntityWithInputFilterInput();
-        $builder = new Annotation\AnnotationBuilder();
-        $form = $builder->createForm($entity, $useAttributes);
-        $inputFilter = $form->getInputFilter();
-
-        $this->assertTrue($inputFilter->has('input'));
-        $expected = [
-            'Laminas\InputFilter\InputInterface',
-            'LaminasTest\Form\TestAsset\Annotation\InputFilterInput',
-        ];
-        foreach ($expected as $expectedInstance) {
-            $this->assertInstanceOf($expectedInstance, $inputFilter->get('input'));
-        }
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     * @group issue-6753
-     */
-    public function testInputFilterAnnotationAllowsComposition($useAttributes)
-    {
-        $entity = new TestAsset\Annotation\EntityWithInputFilterAnnotation();
-        $builder = new Annotation\AnnotationBuilder();
-        $form = $builder->createForm($entity, $useAttributes);
-        $inputFilter = $form->getInputFilter();
-        $this->assertCount(2, $inputFilter->get('username')->getValidatorChain());
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testLegacyComposedObjectAnnotation($useAttributes)
-    {
-        $this->expectDeprecation();
-        $this->expectDeprecationMessageMatches('/Passing a single array .* is deprecated/');
-        $entity  = new TestAsset\Annotation\LegacyComposedObjectAnnotation();
-        $builder = new Annotation\AnnotationBuilder();
-        $form    = $builder->createForm($entity, $useAttributes);
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testLegacyStyleFilterAnnotations($useAttributes)
-    {
-        $this->expectDeprecation();
-        $this->expectDeprecationMessageMatches('/Passing a single array .* is deprecated/');
-        $entity = new TestAsset\Annotation\LegacyFilterAnnotation();
-        $builder = new Annotation\AnnotationBuilder();
-        $form = $builder->createForm($entity, $useAttributes);
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testLegacyStyleHydratorAnnotations($useAttributes)
-    {
-        $this->expectDeprecation();
-        $this->expectDeprecationMessageMatches('/Passing a single array .* is deprecated/');
-        $entity = new TestAsset\Annotation\LegacyHydratorAnnotation();
-        $builder = new Annotation\AnnotationBuilder();
-        $form = $builder->createForm($entity, $useAttributes);
-    }
-
-    /**
-     * @dataProvider useAttributesDataProvider
-     */
-    public function testLegacyStyleValidatorAnnotations($useAttributes)
-    {
-        $this->expectDeprecation();
-        $this->expectDeprecationMessageMatches('/Passing a single array .* is deprecated/');
-        $entity = new TestAsset\Annotation\LegacyValidatorAnnotation();
-        $builder = new Annotation\AnnotationBuilder();
-        $form = $builder->createForm($entity, $useAttributes);
+        return new Annotation\AnnotationBuilder();
     }
 }
