+++ test/Annotation/BuilderAbstractFactoryTest.php
@@ -6,9 +6,12 @@ use Interop\Container\ContainerInterface;
 use Laminas\EventManager\EventManagerInterface;
 use Laminas\EventManager\ListenerAggregateInterface;
 use Laminas\Form\Annotation\AnnotationBuilder;
-use Laminas\Form\Annotation\AnnotationBuilderFactory;
+use Laminas\Form\Annotation\AttributeBuilder;
+use Laminas\Form\Annotation\BuilderAbstractFactory;
+use Laminas\Form\Exception\IncompatiblePhpVersionException;
 use Laminas\Form\FormElementManager;
 use Laminas\ServiceManager\Exception\ServiceNotCreatedException;
+use PhpParser\Node\Attribute;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use ReflectionProperty;
@@ -16,7 +19,7 @@ use stdClass;
 
 use function get_class;
 
-class AnnotationBuilderFactoryTest extends TestCase
+class BuilderAbstractFactoryTest extends TestCase
 {
     use ProphecyTrait;
 
@@ -31,11 +34,70 @@ class AnnotationBuilderFactoryTest extends TestCase
         $container->has('config')->willReturn(false);
         $container->has('InputFilterManager')->willReturn(false);
 
-        $factory = new AnnotationBuilderFactory();
+        $factory = new BuilderAbstractFactory();
+
+        $this->assertTrue($factory->canCreate($container->reveal(), AnnotationBuilder::class));
         $this->assertInstanceOf(
             AnnotationBuilder::class,
             $factory($container->reveal(), AnnotationBuilder::class)
         );
+
+        $this->assertTrue($factory->canCreate($container->reveal(), 'FormAnnotationBuilder'));
+        $this->assertInstanceOf(
+            AnnotationBuilder::class,
+            $factory($container->reveal(), 'FormAnnotationBuilder')
+        );
+    }
+
+    public function testFactoryReturnsAttributeBuilderForPhp8()
+    {
+        if (PHP_MAJOR_VERSION < 8) {
+            $this->markTestSkipped('Can only create attribute builder for PHP >= 8.0.');
+        }
+
+        $container = $this->prophesize(ContainerInterface::class);
+        $events = $this->prophesize(EventManagerInterface::class);
+
+        $elements = $this->prophesize(FormElementManager::class);
+        $container->get('EventManager')->willReturn($events->reveal());
+        $container->get('FormElementManager')->willReturn($elements->reveal());
+        $container->has('config')->willReturn(false);
+        $container->has('InputFilterManager')->willReturn(false);
+
+        $factory = new BuilderAbstractFactory();
+
+        $this->assertTrue($factory->canCreate($container->reveal(), AttributeBuilder::class));
+        $this->assertInstanceOf(
+            AttributeBuilder::class,
+            $factory($container->reveal(), AttributeBuilder::class)
+        );
+
+        $this->assertTrue($factory->canCreate($container->reveal(), 'FormAttributeBuilder'));
+        $this->assertInstanceOf(
+            AttributeBuilder::class,
+            $factory($container->reveal(), 'FormAttributeBuilder')
+        );
+    }
+
+    public function testFactoryReturnsNoAttributeBuilderForPhp7()
+    {
+        if (PHP_MAJOR_VERSION >= 8) {
+            $this->markTestSkipped('Should only throw exceptions when creating attribute builder in PHP < 8.0.');
+        }
+
+        $container = $this->prophesize(ContainerInterface::class);
+        $events = $this->prophesize(EventManagerInterface::class);
+
+        $elements = $this->prophesize(FormElementManager::class);
+        $container->get('EventManager')->willReturn($events->reveal());
+        $container->get('FormElementManager')->willReturn($elements->reveal());
+        $container->has('config')->willReturn(false);
+        $container->has('InputFilterManager')->willReturn(false);
+
+        $factory = new BuilderAbstractFactory();
+
+        $this->expectException(IncompatiblePhpVersionException::class);
+        $factory($container->reveal(), 'FormAttributeBuilder');
     }
 
     public function testFactoryCanSetPreserveDefinedOrderFlagFromConfiguration()
@@ -54,7 +116,7 @@ class AnnotationBuilderFactoryTest extends TestCase
             ],
         ]);
 
-        $factory = new AnnotationBuilderFactory();
+        $factory = new BuilderAbstractFactory();
         $builder = $factory($container->reveal(), AnnotationBuilder::class);
 
         $this->assertTrue($builder->preserveDefinedOrder(), 'Preserve defined order was not set correctly');
@@ -83,7 +145,7 @@ class AnnotationBuilderFactoryTest extends TestCase
         ]);
         $container->get('test-listener')->willReturn($listener->reveal());
 
-        $factory = new AnnotationBuilderFactory();
+        $factory = new BuilderAbstractFactory();
         $factory($container->reveal(), AnnotationBuilder::class);
     }
 
@@ -108,7 +170,7 @@ class AnnotationBuilderFactoryTest extends TestCase
         ]);
         $container->get('test-listener')->willReturn($listener->reveal());
 
-        $factory = new AnnotationBuilderFactory();
+        $factory = new BuilderAbstractFactory();
 
         $this->expectException(ServiceNotCreatedException::class);
         $this->expectExceptionMessage('Invalid event listener');
